datafarmer/pyproject.toml
[tool.poetry]
name = "datafarmer"
version = "0.2.0"
description = "this the simple package for a data farmer. wrapping several complex task in to simple"
authors = ["Prasetyo Wianto <prasetyo.wianto@traveloka.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
pytest = "^8.3.3"
pandas = "^2.2.2"
google-cloud-bigquery = "^3.25.0"
db-dtypes = "^1.3.0"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scipy = "^1.14.1"
google-cloud-aiplatform = "^1.67.1"
polars = "^1.7.1"
tenacity = "^9.0.0"
tqdm = "^4.66.5"
typer = "^0.12.5"
grpcio = "1.60"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
datafarmer="cli.main:app"
